fix(frontend/CreateProductForm): Accept decimals in any format

+++ frontend/src/products/views/CreateProductView.tsx
@@ -5,6 +5,7 @@ import { ViewContainer } from "../../common/components/ViewContainer";
 import { CreateProductForm } from "../containers/CreateProductForm";
 import { createProduct } from "../services/ProductsService";
 import { useRefetchProducts } from "../hooks/useRefetchProducts";
+import { parseDecimal } from "../../common/services/decimalNumbers";
 
 export function CreateProductView() {
   const navigate = useNavigate();
@@ -17,7 +18,7 @@ export function CreateProductView() {
           createProduct({
             measurement_unit_id: parseInt(measurement_unit_id),
             product_name,
-            list_unit_price: parseFloat(price),
+            list_unit_price: parseDecimal(price),
           }).then(() => {
             refetchProducts();
             navigate("/products");
